+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - add @ApiOptional to mark optional parameters
+- parse javax.validation constraints as allowedValueHint
 
 ### Changed